requirements.txt
browser-cookie3==0.11.1
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
decorator==4.4.2
firebase-admin==4.3.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-WTF==0.14.3
gcloud==0.17.0
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-cloud-core==1.4.1
google-cloud-firestore==1.8.1
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
grpcio==1.31.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jws==0.1.3
keyring==21.3.0
lz4==3.1.0
MarkupSafe==1.1.1
moviepy==1.0.3
msgpack==1.0.0
numpy==1.19.1
oauth2client==3.0.0
pbkdf2==1.3
Pillow==7.2.0
proglog==0.1.9
protobuf==4.0.0rc2
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
PyJWT==2.0.1
Pyrebase
python-dotenv==0.14.0
python-jwt==2.0.1
pytube @ git+https://github.com/huishun98/pytube.git
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.7.0
rsa==4.6
six==1.15.0
tqdm==4.48.2
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.1.0
